refactor: move backend mock podcast statistic data to data directory
