+++ backend/__tests__/mocks/data/podcastStatistic.ts
@@ -1,0 +1,14 @@
+// https://podcastindex-org.github.io/docs-api/#get-/stats/current
+export const CURRENT_PODCAST_STATISTICS = {
+  status: "true",
+  stats: {
+    feedCountTotal: 4257865,
+    episodeCountTotal: 143953477,
+    feedsWithNewEpisodes3days: 97522,
+    feedsWithNewEpisodes10days: 212322,
+    feedsWithNewEpisodes30days: 306748,
+    feedsWithNewEpisodes90days: 436933,
+    feedsWithValueBlocks: 21916,
+  },
+  description: "Found matching feed",
+}
